+++ Changelog.md
@@ -1,5 +1,6 @@
 # To Do's
 
-* POST
-* PUT
 * Pictures
+* modify db. add `order`.`in_progress`
+* ~~Change DB to default to "Germany" when entering an address~~
+* ~~build endpoint type endpoint for API at `GET /endpointType`~~